Fix integer overflow in size computations across VP8/VP9 encoder
Harden 30+ integer overflow sites in pixel-dimension arithmetic,
buffer allocation size calculations, and PSNR metric computations.

The root cause is that cm->width * cm->height (both int) and similar
dimension-derived multiplications are performed in 32-bit signed
arithmetic throughout the codebase, despite VP9 supporting dimensions
up to 65536x65536. This creates undefined behavior (signed overflow),
incorrect allocation sizes, and a division-by-zero in rate control.

This is the same class of vulnerability that led to CVE-2023-5217
(heap buffer overflow in VP8 encoding, exploited as a 0-day) and
CVE-2024-5197 (integer overflow in vpx_img_alloc).

Changes:
- Add vpx_size_checked_mul() overflow-checked helpers (vpx_mem.h)
- Fix VP8 vp8_yv12_realloc_frame_buffer to use uint64_t like VP9's
- Fix get_token_alloc() int return type overflow (int -> int64_t)
- Fix cx_data_sz overflow in VP8 and VP9 encoder init
- Fix VP8 token allocation overflow in onyx_if.c
- Fix PSNR uint32_t samples overflow in vpx_dsp/psnr.c
- Fix 23 cm->width * cm->height signed overflow sites across 6 files

Bug: b/none

+++ vp8/vp8_cx_iface.c
@@ -89,7 +89,7 @@ struct vpx_codec_alg_priv {
   VP8_CONFIG oxcf;
   struct VP8_COMP *cpi;
   unsigned char *cx_data;
-  unsigned int cx_data_sz;
+  size_t cx_data_sz;
   vpx_image_t preview_img;
   unsigned int next_frame_flag;
   vp8_postproc_cfg_t preview_ppcfg;
@@ -713,7 +713,7 @@ static vpx_codec_err_t vp8e_init(vpx_codec_ctx_t *ctx,
     priv->vp8_cfg = default_extracfg;
     priv->vp8_cfg.pkt_list = &priv->pkt_list.head;
 
-    priv->cx_data_sz = priv->cfg.g_w * priv->cfg.g_h * 3 / 2 * 2;
+    priv->cx_data_sz = (size_t)priv->cfg.g_w * priv->cfg.g_h * 3 / 2 * 2;
 
     if (priv->cx_data_sz < 32768) priv->cx_data_sz = 32768;
